requirements.txt
pandas
BioPython
openpyxl